Refactorized Gnap requestAccess api, separated from initLocalStorage()
